PLT-3313: Always use `SemiArbitrary`
