+++ marlowe-spec-test/src/Marlowe/Spec/Core/Serialization/Json.hs
@@ -15,33 +15,28 @@ import qualified Data.Aeson.Types as JSON
 import Data.Proxy (Proxy (..))
 import Data.Text as T
 import GHC.Stack (HasCallStack)
-import Marlowe.Spec.Core.Generators
-  ( genAction,
-    genChoiceId,
-    genContract,
-    genInput,
-    genObservation,
-    genParty,
-    genPayee,
-    genPayment,
-    genState,
-    genToken,
-    genTransaction,
-    genTransactionOutput,
-    genTransactionWarning,
-    genValue,
+import Marlowe.Spec.Core.Generators (gen)
+import Marlowe.Spec.Interpret
+  ( InterpretJsonRequest,
+    Request (..),
+    Response (..),
+    testResponse,
   )
-import Marlowe.Spec.Interpret (InterpretJsonRequest, Request (..), Response (..), testResponse)
-import Marlowe.Spec.Reproducible (assertResponse, generate, generateT, reproducibleProperty)
+import Marlowe.Spec.Reproducible
+  ( assertResponse,
+    generate,
+    generateT,
+    reproducibleProperty,
+  )
 import Marlowe.Spec.TypeId (HasTypeId (..), TypeId (..))
 import MarloweCoreJson
-import QuickCheck.GenT (MonadGen (resize))
+import QuickCheck.GenT (GenT, MonadGen (resize))
 import qualified Semantics as C
 import qualified SemanticsTypes as C
+import Test.QuickCheck (Arbitrary (..), Gen)
 import Test.QuickCheck.Monadic (PropertyM, run)
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit (Assertion, assertBool, testCase, (@=?))
-import Test.QuickCheck (Gen, Arbitrary (..))
 
 data SerializationResponse transport
   = SerializationSuccess transport
@@ -182,23 +177,23 @@ isUnknownType _ = False
 arbitraryTokenTest :: InterpretJsonRequest -> TestTree
 arbitraryTokenTest i = reproducibleProperty "Token" do
   -- Any token that is randomly generated by the interpreter should also pass the roundtrip test
-  token <- run $ generateT $ genToken i
+  token <- run $ generateT (gen i :: GenT IO C.Token)
   propertyRoundtripTest i token
 
 arbitraryPartyTest :: InterpretJsonRequest -> TestTree
 arbitraryPartyTest i = reproducibleProperty "Party" do
   -- Any party that is randomly generated by the interpreter should also pass the roundtrip test
-  party <- run $ generateT $ genParty i
+  party <- run $ generateT (gen i :: GenT IO C.Party)
   propertyRoundtripTest i party
 
 arbitraryPayeeTest :: InterpretJsonRequest -> TestTree
 arbitraryPayeeTest i = reproducibleProperty "Payee" do
-  payee <- run $ generateT $ genPayee i
+  payee <- run $ generateT (gen i :: GenT IO C.Payee)
   propertyRoundtripTest i payee
 
 arbitraryChoiceIdTest :: InterpretJsonRequest -> TestTree
 arbitraryChoiceIdTest i = reproducibleProperty "ChoiceId" do
-  choiceId <- run $ generateT $ genChoiceId i
+  choiceId <- run $ generateT (gen i :: GenT IO C.ChoiceId)
   propertyRoundtripTest i choiceId
 
 arbitraryBoundTest :: InterpretJsonRequest -> TestTree
@@ -208,47 +203,47 @@ arbitraryBoundTest i = reproducibleProperty "Bound" do
 
 arbitraryValueTest :: InterpretJsonRequest -> TestTree
 arbitraryValueTest i = reproducibleProperty "Value" do
-  value <- run $ generateT $ resize 15 $ genValue i
+  value <- run $ generateT $ resize 15 (gen i :: GenT IO C.Value)
   propertyRoundtripTest i value
 
 arbitraryObservationTest :: InterpretJsonRequest -> TestTree
 arbitraryObservationTest i = reproducibleProperty "Observation" do
-  observation <- run $ generateT $ resize 15 $ genObservation i
+  observation <- run $ generateT $ resize 15 (gen i :: GenT IO C.Observation)
   propertyRoundtripTest i observation
 
 arbitraryActionTest :: InterpretJsonRequest -> TestTree
 arbitraryActionTest i = reproducibleProperty "Action" do
-  action <- run $ generateT $ resize 15 $ genAction i
+  action <- run $ generateT $ resize 15 (gen i :: GenT IO C.Action)
   propertyRoundtripTest i action
 
 arbitraryContractTest :: InterpretJsonRequest -> TestTree
 arbitraryContractTest i = reproducibleProperty "Contract" do
-  contract <- run $ generateT $ resize 10 $ genContract i
+  contract <- run $ generateT $ resize 10 (gen i :: GenT IO C.Contract)
   propertyRoundtripTest i contract
 
 arbitraryInputTest :: InterpretJsonRequest -> TestTree
 arbitraryInputTest i = reproducibleProperty "Input" do
-  input <- run $ generateT $ genInput i
+  input <- run $ generateT (gen i :: GenT IO C.Input)
   propertyRoundtripTest i input
 
 arbitraryTransactionTest :: InterpretJsonRequest -> TestTree
 arbitraryTransactionTest i = reproducibleProperty "Transaction" do
-  tx <- run $ generateT $ genTransaction i
+  tx <- run $ generateT (gen i :: GenT IO (C.Transaction_ext ()))
   propertyRoundtripTest i tx
 
 arbitraryPaymentTest :: InterpretJsonRequest -> TestTree
 arbitraryPaymentTest i = reproducibleProperty "Payment" do
-    payment <- run $ generateT $ genPayment i
+    payment <- run $ generateT (gen i :: GenT IO C.Payment)
     propertyRoundtripTest i payment
 
 arbitraryStateTest :: InterpretJsonRequest -> TestTree
 arbitraryStateTest i = reproducibleProperty "State" do
-  state <- run $ generateT $ genState i
+  state <- run $ generateT (gen i :: GenT IO (C.State_ext ()))
   propertyRoundtripTest i state
 
 arbitraryTransactionWarningTest :: InterpretJsonRequest -> TestTree
 arbitraryTransactionWarningTest i = reproducibleProperty "TransactionWarning" do
-  warning <- run $ generateT $ genTransactionWarning i
+  warning <- run $ generateT (gen i :: GenT IO C.TransactionWarning)
   propertyRoundtripTest i warning
 
 arbitraryIntervalErrorTest :: InterpretJsonRequest -> TestTree
@@ -263,5 +258,5 @@ arbitraryTransactionErrorTest i = reproducibleProperty "TransactionError" do
 
 arbitraryTransactionOutputTest :: InterpretJsonRequest -> TestTree
 arbitraryTransactionOutputTest i = reproducibleProperty "TransactionOutput" do
-  out <- run $ generateT $ genTransactionOutput i
+  out <- run $ generateT (gen i :: GenT IO C.TransactionOutput)
   propertyRoundtripTest i out
